refactor : 북마크 컴포넌트 네이밍 수정
